Change add to cart method to use ajax via Selenium
Request transfer of cookies to Chrome did not always yield a functioning cart. Often times it would result in an infinite redirection-loop which would cause soft-bans.  The avoid this issue the approach addToCartRequestTransferToChrome is ditched in favor of addToCartChromeAJAX which uses ajax.  This approach allows us to avoid CORS.

If the request transfer can be improved to minimize the infinite redirection-loop the add-to-cart mechanism might change in the future. 

Session cookies are now deleted at the start and end of each webdriver session.